feat : add monthsalary and age calculation

+++ src/main/java/com/example/prog4/model/Employee.java
@@ -11,6 +11,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -35,7 +36,7 @@ public class Employee implements Serializable {
     private String personalEmail;
     private String professionalEmail;
     private String registrationNumber;
-    private double monthlySalary;
+    private BigDecimal monthlySalary;
     private int age;
 
     @DateTimeFormat(pattern="yyyy-MM-dd")
